+++ DAL/DAL_ChiTietXuatKho.cs
@@ -23,14 +23,14 @@ namespace DAL
         }
 
         //Thêm chi tiết xuất kho
-        public void AddExportDetail(string maphieuxuat, string manl, int soluong)
+        public void AddExportDetail(string maphieuxuat, string manl, decimal soluong)
         {
             string query = "insert into ChiTietXuatKho values (@_MaPhieuXuat,@_MaNL,@_SoLuong)";
             object[] parem = new object[] { maphieuxuat, manl, soluong };
             DataProvider.ExecuteNonQuery(query, parem);
         }
 
-        public void UpdateDetailst(string maphieuxuat, string manl, int soluong)
+        public void UpdateDetailst(string maphieuxuat, string manl, decimal soluong)
         {
             string query = "update ChiTietXuatKho set manl = dbo.LayMaNlTheoTenNl(@_MaNL), soluong = @_SoLuong where maphieuxuat = @_MaPhieuXuat";
             object[] parem = new object[] { manl, soluong, maphieuxuat };
